Extract DataprocHub server in its own class.
